renamed glossary/initial_data.yaml to starter_glossary.yaml
Django is designed so that initial_data.* gets reloaded each time that
syncdb is run:*
http://docs.djangoproject.com/en/dev/howto/initial-data/#providing-initial-data-with-fixtures

Since I did not include primary keys in the file, the fixture data gets
appended each time.  This is not what we want.

* For what it is worth, there is a thread on the Django developers list
that argues against the way Django loads initial_data.*:
http://groups.google.com/group/django-developers/browse_thread/thread/9e3444a7afebd579

In any case, the easiest solution for now is to rename the file.  This
means it has to be loaded explicitly using this command:
 `./manage.py loaddata starter_glossary`